Add abstraction layer for posix
